Добавлены звуковые сигналы. Реализовано скачивание текстов песен. И ещё некоторые изменения.

+++ ymd-console-client/src/main/kotlin/ru/nnedition/ymdownloader/MusicDownloader.kt
@@ -6,16 +6,20 @@ import ru.nnedition.ymdownloader.api.download.AbstractMusicDownloader
 import ru.nnedition.ymdownloader.api.ffmpeg.FfmpegProvider
 import ru.nnedition.ymdownloader.api.link.LinkInfo
 import ru.nnedition.ymdownloader.api.link.LinkType
+import ru.nnedition.ymdownloader.api.objects.LyricInfo
 import ru.nnedition.ymdownloader.api.objects.Track
 import ru.nnedition.ymdownloader.api.objects.album.Album
 import ru.nnedition.ymdownloader.api.objects.artist.ArtistMetaResult
 import ru.nnedition.ymdownloader.api.utils.AudioTagWriter
 import ru.nnedition.ymdownloader.api.utils.GenreTranslator
 import ru.nnedition.ymdownloader.api.utils.createDir
+import ru.nnedition.ymdownloader.audio.AudioPlayer
+import ru.nnedition.ymdownloader.audio.AudioType
 import ru.nnedition.ymdownloader.terminal.context.GenreSelectContext
 import ru.nnedition.ymdownloader.terminal.context.GenreTranslateContext
 import java.io.File
 import java.lang.Thread.currentThread
+import java.util.concurrent.Executors
 import java.util.concurrent.LinkedBlockingQueue
 import kotlin.concurrent.thread
 
@@ -28,7 +32,8 @@ class MusicDownloader(
     @Volatile
     var paused = false
 
-    private val queue = LinkedBlockingQueue<DownloadItem>()
+    val toDownloadQueue = LinkedBlockingQueue<DownloadItem>()
+    val downloadQueue = LinkedBlockingQueue<DownloadItem>()
 
     val thread = thread(
         isDaemon = false,
@@ -36,7 +41,7 @@ class MusicDownloader(
         priority = 1
     ) {
         while (!currentThread().isInterrupted) {
-            val item = this.queue.take()
+            val item = this.downloadQueue.take()
             @Suppress("ControlFlowWithEmptyBody")
             while (this.paused) {}
 
@@ -67,11 +72,20 @@ class MusicDownloader(
 
             val toMuxFile = File(path, "$fileName.mp4")
 
-            try {
-                downloadTrack(info, if (shouldMux) toMuxFile else finalFile)
-            } catch (e: Exception) {
-                toMuxFile.delete()
-                e.printStackTrace()
+            var successDownload = false
+            repeat(3) { attempt ->
+                try {
+                    downloadTrack(info, if (shouldMux) toMuxFile else finalFile)
+                    successDownload = true
+                    return@repeat
+                } catch (e: Exception) {
+                    logger.error("Ошибка при ${attempt+1} попытке загрузки трека ${track.fullTitle}: ${e.message}")
+                }
+            }
+
+            if (!successDownload) {
+                if (shouldMux)
+                    toMuxFile.delete()
                 continue
             }
 
@@ -101,46 +115,50 @@ class MusicDownloader(
                 finalFile = interimFile
             }
 
-            @Suppress("ControlFlowWithEmptyBody")
-            val genre = (track.genre ?: track.album.genre)?.let { genre ->
-                GenreTranslator.translate(genre) ?: let {
-                    logger.warn("Найден неизвестный жанр: \"$genre\"")
-                    Launcher.terminal.context = GenreTranslateContext(genre)
-                    while (GenreTranslator.translate(genre) == null) {}
-                    GenreTranslator.translate(genre)
-                }
-            } ?: run {
-                logger.warn("Жанр трека \"${track.title}\" не найден. Введите его вручную:")
-                val context = GenreSelectContext()
-                Launcher.terminal.context = context
-                while (context.genre == null) {}
-                context.genre!!
-            }
-
             AudioTagWriter.write(
                 file = finalFile,
                 track = track,
-                genre = genre,
+                genre = item.genre,
                 cover = this.ymClient.getCoverData(track.coverUri, false, false, "300x300"),
             )
+
+            item.lyricInfo?.let { lyricInfo ->
+                File(path, "${finalFile.nameWithoutExtension}.lrc").takeIf { !it.exists() }?.let { lyricFile ->
+                    downloadLyric(lyricInfo, lyricFile)
+                }
+            }
+
+            println("Загрузка ${if (track.album.isSingle()) "сингла" else "трека" } \"${track.fullTitle}\" окончена.")
+
+            if (this.downloadQueue.isEmpty())
+                println("Загрузка завершена.")
         }
     }
 
-    fun download(info: LinkInfo) {
-        runCatching {
-            when (info.type) {
-                LinkType.ARTIST -> downloadArtist(info.id)
-                LinkType.ALBUM -> downloadAlbum(info.id)
-                LinkType.TRACK -> downloadTrack(info.id)
-            }
+    val processExecutor = Executors.newSingleThreadExecutor()
 
-            println("Скачивание завершено.")
-        }.onFailure { it.printStackTrace() }
+    fun download(info: LinkInfo) {
+        this.processExecutor.execute {
+            runCatching {
+                when (info.type) {
+                    LinkType.ARTIST -> downloadArtist(info.id)
+                    LinkType.ALBUM -> downloadAlbum(info.id)
+                    LinkType.TRACK -> downloadTrack(info.id)
+                }
+
+                Launcher.terminal.terminal.writer().also {
+                    it.print("\r\u001B[2K")
+                    it.print("Обработка треков окончена. Всего добавлено в очередь: ${this.toDownloadQueue.size}.")
+                    it.flush()
+                }
+
+                this.downloadQueue.addAll(this.toDownloadQueue)
+                this.toDownloadQueue.clear()
+            }.onFailure { it.printStackTrace() }
+        }
     }
 
     override fun downloadArtist(artist: ArtistMetaResult, config: IConfiguration) {
-        logger.info("Обработка артиста \"${artist.artist.name}\"...")
-
         artist.albums.forEach { album ->
             downloadAlbum(album.id, config)
         }
@@ -148,11 +166,9 @@ class MusicDownloader(
 
     override fun downloadAlbum(album: Album, config: IConfiguration) {
         if (!album.available) {
-            logger.error("${if (album.isSingle()) "Сингл" else "Альбом"} \"${album.title}\" недоступен")
+            logger.error("${if (album.isSingle()) "Сингл" else "Альбом"} \"${album.title}\" (${album.publisher.name}) недоступен")
             return
         }
-
-        logger.info("Обработка ${if (album.isSingle()) "сингла" else "альбома" } \"${album.title}\"...")
 
         album.tracks.forEach { track ->
             downloadTrack(track, config)
@@ -161,20 +177,52 @@ class MusicDownloader(
 
     override fun downloadTrack(track: Track, config: IConfiguration) {
         if (!track.available) {
-            logger.error("Трек \"${track.fullTitle}\" недоступен")
+            logger.error("Трек \"${track.fullTitle}\" (${track.publisher.name} - ${track.album.title}) недоступен")
             return
         }
 
         track.album = this.ymClient.getAlbum(track.albums[0].id)
 
-        if (!track.album.isSingle())
-            logger.info("Обработка трека \"${track.fullTitle}\"...")
+        @Suppress("ControlFlowWithEmptyBody")
+        val genre = (track.genre ?: track.album.genre)?.let { genre ->
+            GenreTranslator.translate(genre) ?: let {
+                Launcher.terminal.context = GenreTranslateContext(genre)
+                logger.warn("Найден неизвестный жанр: \"$genre\"")
+                AudioPlayer.play(AudioType.NOTIFICATION)
+                while (GenreTranslator.translate(genre) == null) {}
+                GenreTranslator.translate(genre)
+            }
+        } ?: run {
+            val context = GenreSelectContext()
+            Launcher.terminal.context = context
+            logger.warn("Жанр трека \"${track.title}\" не найден. Введите его вручную:")
+            AudioPlayer.play(AudioType.NOTIFICATION)
+            while (context.genre == null) {}
+            context.genre!!
+        }
 
-        this.queue.put(DownloadItem(config, track))
+        val item = DownloadItem(
+            config,
+            track,
+            genre,
+            ymClient.getLyricInfo(track.id.toString()),
+        )
+
+        this.toDownloadQueue.put(item)
+
+        for ((index, track) in this.toDownloadQueue.withIndex()) {
+            Launcher.terminal.terminal.writer().also {
+                it.print("\r\u001B[2K")
+                it.print("Обработка треков: ${track.track.fullTitle} (${index + 1})")
+                it.flush()
+            }
+        }
     }
 
     class DownloadItem(
         val config: IConfiguration,
         val track: Track,
+        val genre: String,
+        val lyricInfo: LyricInfo?,
     )
 }
